refactor: Move components to named files
I was getting sick of seeing a bunch of "index.js" files in my editor.
