+++ src/components/Banner/Banner.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import styles from './Banner.module.scss';
+
+const Banner = ({children}) => (
+    <div className={styles.banner}>
+        <div className={styles.container}>{children}</div>
+    </div>
+);
+
+export default Banner;
