+++ src/components/CallToAction/index.js
@@ -1,11 +1,3 @@
-import React from 'react';
-
-import styles from './index.module.scss';
-
-const CallToAction = props => (
-    <div className={styles.cta}>
-        <div className={styles.container}>{props.children}</div>
-    </div>
-);
+import CallToAction from './CallToAction';
 
 export default CallToAction;
