+++ src/components/Container/Container.js
@@ -1,0 +1,18 @@
+import React from 'react';
+import classNames from 'classnames';
+
+import styles from './Container.module.scss';
+
+const Container = ({children, isNarrow, className, ...props}) => (
+    <div
+        className={classNames({
+            [styles.container]: true,
+            [className]: true,
+        })}
+        {...props}
+    >
+        {isNarrow ? <div className={styles.narrow}>{children}</div> : children}
+    </div>
+);
+
+export default Container;
