+++ src/components/Container/index.js
@@ -1,18 +1,3 @@
-import React from 'react';
-import classNames from 'classnames';
-
-import styles from './index.module.scss';
-
-const Container = ({children, isNarrow, className, ...props}) => (
-    <div
-        className={classNames({
-            [styles.container]: true,
-            [className]: true,
-        })}
-        {...props}
-    >
-        {isNarrow ? <div className={styles.narrow}>{children}</div> : children}
-    </div>
-);
+import Container from './Container';
 
 export default Container;
