+++ src/components/DocsContainer/index.js
@@ -1,17 +1,3 @@
-import React from 'react';
-
-import Container from '../Container';
-import styles from './index.module.scss';
-
-const DocsContainer = ({children, TOC}) => (
-    <Container className={styles.container}>
-        <nav className={styles.nav}>
-            <TOC />
-        </nav>
-        <div className={styles.content}>
-            {children}
-        </div>
-    </Container>
-);
+import DocsContainer from './DocsContainer';
 
 export default DocsContainer;
