+++ src/components/Faq/index.js
@@ -1,12 +1,3 @@
-import React from 'react';
-
-import styles from './index.module.scss';
-
-const Faq = ({title, children}) => (
-    <React.Fragment>
-        <h3 className={styles.title}>{title}</h3>
-        <div className={styles.body}>{children}</div>
-    </React.Fragment>
-);
+import Faq from './Faq';
 
 export default Faq;
