+++ src/components/LinkButton/index.js
@@ -1,16 +1,3 @@
-import React from 'react';
-import classNames from 'classnames';
-
-import styles from './index.module.scss';
-
-const LinkButton = ({className, ...props}) => (
-    <button
-        className={classNames({
-            [styles.linkbutton]: true,
-            [className]: true,
-        })}
-        {...props}
-    />
-);
+import LinkButton from './LinkButton';
 
 export default LinkButton;
